setting up structure
